seyoung

+++ src/main/java/com/mentaljava/crewsjpatest/repository/CrewRepository.java
@@ -3,6 +3,6 @@ package com.mentaljava.crewsjpatest.repository;
 import com.mentaljava.crewsjpatest.domain.Crew;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CrewRepository extends JpaRepository<Crew,Long> {
+public interface CrewRepository extends JpaRepository<Crew,Integer> {
 
 }
